+++ api/ghc/[...path].ts
@@ -1,0 +1,5 @@
+import { handleVercelRequest } from "../../src/vercel";
+
+export default {
+  fetch: handleVercelRequest,
+};